Quick backup for files/folders/programs (Linux)
This program allows you to quickly make a backup of a file/folder/program, which is then hashed (256) and checked against the original, then stored in the "Backups" folder in the same location as the backup was made, for ease of locating if the need arises.

History: This program came about as we were, once again, doing website editing, editing php files, index files galore, and were getting more and more beautiful results. We wanted to save all of them, so we would do the whole "sudo cp file.txt file_MMDDYYY_HHMM.txt thing and typing all of that took...time. I'm no amanuensis, forgive me, though I don't natively type that fast...yet. We needed something that could make a backup by just typing "backup file.txt", and after we got it, we realized, we need to make sure it was a pure backup, so we added Sha256sum verification by comparing the sha's of the original and the backup. Annnnnnd we got it. And then we realized, we probably need to be able to backup an entire folder. Annnnnnnnnd we got it! And now we're sharing it with you. Yay for you! IF you notice anything thing worth noticing, message us, my friend!!! -Mr.GHrey@HellsingAcademy.com

Though it was made on Linux, you can attempt to run it on a different OS and just install what ever binaries or files you need. It'll tell you in the error codes if it doesn't play well, initially.